fake-news-detector-v0/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.7
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
cycler==0.10.0
flatbuffers==1.12
gast==0.4.0
gensim==3.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.3
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyMySQL==1.0.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
selenium==3.141.0
six==1.15.0
smart-open==5.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.62.0
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
Flask>=1.0.0
Flask-SQLAlchemy==2.3.2
google-cloud-datastore==1.7.1
gunicorn==19.9.0
sklearn